# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IFilmCover": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"film_id": {
						"type": "number",
						"format": "double"
					},
					"cover_path": {
						"type": "string"
					},
					"file_type": {
						"type": "string"
					},
					"file_size": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"film_id",
					"cover_path",
					"file_type",
					"file_size",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Film": {
				"properties": {
					"film_id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"release_year": {
						"type": "number",
						"format": "double"
					},
					"language_id": {
						"type": "number",
						"format": "double"
					},
					"original_language_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"rental_duration": {
						"type": "number",
						"format": "double"
					},
					"rental_rate": {
						"type": "number",
						"format": "double"
					},
					"length": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"replacement_cost": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "string"
					},
					"special_features": {
						"type": "string",
						"nullable": true
					},
					"last_update": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"film_id",
					"title",
					"description",
					"release_year",
					"language_id",
					"original_language_id",
					"rental_duration",
					"rental_rate",
					"length",
					"replacement_cost",
					"rating",
					"special_features",
					"last_update"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedFilms": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Film"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"pageSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FilmCreationParams": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"release_year": {
						"type": "number",
						"format": "double"
					},
					"language_id": {
						"type": "number",
						"format": "double"
					},
					"original_language_id": {
						"type": "number",
						"format": "double"
					},
					"rental_duration": {
						"type": "number",
						"format": "double"
					},
					"rental_rate": {
						"type": "number",
						"format": "double"
					},
					"length": {
						"type": "number",
						"format": "double"
					},
					"replacement_cost": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "string"
					},
					"special_features": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"release_year",
					"language_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_FilmCreationParams_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"release_year": {
						"type": "number",
						"format": "double"
					},
					"language_id": {
						"type": "number",
						"format": "double"
					},
					"original_language_id": {
						"type": "number",
						"format": "double"
					},
					"rental_duration": {
						"type": "number",
						"format": "double"
					},
					"rental_rate": {
						"type": "number",
						"format": "double"
					},
					"length": {
						"type": "number",
						"format": "double"
					},
					"replacement_cost": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "string"
					},
					"special_features": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "joel",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "joel"
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"paths": {
...
		"/films/{filmId}/cover": {
			"post": {
				"operationId": "UploadCover",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFilmCover"
								}
							}
						}
					}
				},
				"description": "Télécharger une image de couverture pour un film",
				"tags": [
					"Film Covers"
				],
				"security": [
					{
						"jwt": [
							"create:film"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCover",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IFilmCover"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Récupérer l'image de couverture d'un film",
				"tags": [
					"Film Covers"
				],
				"security": [
					{
						"jwt": [
							"read:films"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCover",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Supprimer l'image de couverture d'un film",
				"tags": [
					"Film Covers"
				],
				"security": [
					{
						"jwt": [
							"delete:film"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/films": {
			"get": {
				"operationId": "GetFilms",
				"responses": {
					"200": {
						"description": "Liste des films récupérée avec succès",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedFilms"
								}
							}
						}
					}
				},
				"description": "Récupère la liste de tous les films avec pagination",
				"tags": [
					"Films"
				],
				"security": [],
				"parameters": [
					{
						"description": "Numéro de la page (commence à 1)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Nombre d'éléments par page",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateFilm",
				"responses": {
					"201": {
						"description": "Film créé avec succès",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Film"
								}
							}
						}
					}
				},
				"description": "Crée un nouveau film",
				"tags": [
					"Films"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Données du film à créer",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FilmCreationParams",
								"description": "Données du film à créer"
							}
						}
					}
				}
			}
		},
		"/films/{id}": {
			"get": {
				"operationId": "GetFilmById",
				"responses": {
					"200": {
						"description": "Film récupéré avec succès",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Film"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Récupère un film par son identifiant",
				"tags": [
					"Films"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identifiant unique du film",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateFilm",
				"responses": {
					"200": {
						"description": "Film mis à jour avec succès",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Film"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Met à jour un film existant",
				"tags": [
					"Films"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Identifiant du film à mettre à jour",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Nouvelles données du film",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_FilmCreationParams_",
								"description": "Nouvelles données du film"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFilm",
				"responses": {
					"204": {
						"description": "Film supprimé avec succès",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Supprime un film",
				"tags": [
					"Films"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Identifiant du film à supprimer",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/films/category/{categoryId}": {
			"get": {
				"operationId": "GetFilmsByCategory",
				"responses": {
					"200": {
						"description": "Films par catégorie récupérés avec succès",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedFilms"
								}
							}
						}
					}
				},
				"description": "Récupère les films par catégorie",
				"tags": [
					"Films"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identifiant de la catégorie",
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Numéro de la page",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Nombre d'éléments par page",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/films/actor/{actorId}": {
			"get": {
				"operationId": "GetFilmsByActor",
				"responses": {
					"200": {
						"description": "Films par acteur récupérés avec succès",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedFilms"
								}
							}
						}
					}
				},
				"description": "Récupère les films par acteur",
				"tags": [
					"Films"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identifiant de l'acteur",
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Numéro de la page",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Nombre d'éléments par page",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/actors/{id}/films": {
			"get": {
				"operationId": "GetActorFilms",
				"responses": {
					"200": {
						"description": "Films de l'acteur récupérés avec succès",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pageSize": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"pageSize",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Récupère les films d'un acteur spécifique",
				"tags": [
					"Actors"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identifiant de l'acteur",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Numéro de la page",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Nombre d'éléments par page",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	}
}